Added action in home btns
